Fix scheduled job repeat working
